move run_record_and_transcribe to own function
